MLModel/requirements.txt
pandas
scikit-learn
python-dotenv
nltk
Flask
